requirements.txt
fastapi
uvicorn
playwright
python-dotenv
pydantic
beautifulsoup4
supabase
requests
google-genai
python-multipart
google-cloud-storage
